requirements.txt
certifi==2019.3.9
chardet==3.0.4
clarifai==2.6.0
Click==7.0
configparser==3.7.3
Flask==1.0.2
future==0.17.1
googleapis-common-protos==1.5.8
grpcio==1.19.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.1.1
numpy==1.16.2
pandas==0.24.1
Pillow==5.4.1
protobuf==3.6.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
protobuf==3.6.1